fix the failing tests on build
